requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.6.3
asgiref==3.7.2
attrs==23.2.0
Authlib==1.2.1
azure-common==1.1.28
azure-core==1.30.0
azure-identity==1.15.0
azure-mgmt-core==1.4.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.0.1
azure-mgmt-subscription==3.1.1
Babel==2.14.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.7.0
blis==0.7.11
boto3==1.33.13
botocore==1.33.13
Brotli==1.1.0
build==1.2.1
cachetools==4.2.2
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
cryptography==41.0.7
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
Django==5.0.7
django-ckeditor==6.7.1
django-cors-headers==4.9.0
django-js-asset==2.2.0
django-modeltranslation==0.19.5
django-phonenumber-field==7.3.0
djangorestframework==3.15.2
dnspython==2.6.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
easy-colors==0.0.1
email-validator==2.1.0.post1
emoji==2.14.0
eval_type_backport==0.2.0
eventlet==0.34.2
facebook-sdk==3.1.0
fastai==2.7.15
fastapi==0.115.5
fastapi-cli==0.0.4
fastcore==1.5.55
fastdownload==0.0.7
fastprogress==1.0.3
fest==5.2.0
ffmpy==0.4.0
filelock==3.13.1
filetype==1.2.0
Flask==2.3.3
flask-babel==4.0.0
Flask-Compress==1.14
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.2.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.62.0
googletrans==4.0.0rc1
gradio==5.6.0
gradio_client==1.4.3
greenlet==3.0.3
grpcio==1.65.1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
html5lib==1.1
httpagentparser==1.9.5
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib-resources==6.1.1
importlib_metadata==8.0.0
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.3.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==24.3.0
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.24
langchain-text-splitters==0.2.2
langcodes==3.4.0
langdetect==1.0.9
langsmith==0.1.93
language_data==1.2.0
ldap3==2.9.1
libretranslatepy==2.1.1
logger==1.4
lxml==5.2.2
Mako==1.3.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
mdurl==0.1.2
meta==1.0.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msal==1.26.0
msal-extensions==1.1.0
msrest==0.7.1
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
mysqlclient==2.2.7
nest-asyncio==1.6.0
networkx==3.3
ngrok==1.3.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.18.1
openai==1.37.1
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.6
overrides==7.7.0
package-json==0.0.0
packaging==23.2
pandas==2.2.2
paramiko==3.4.0
passlib==1.7.4
pbr==6.0.0
pgadmin4==8.3
phone==0.4.5
phonenumbers==8.13.35
Pillow==10.1.0
platformdirs==4.1.0
polib==1.2.0
portalocker==2.8.2
postgres==4.0
posthog==3.5.0
preshed==3.0.9
protobuf==4.25.3
psutil==5.9.8
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg2-pool==1.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.1.1
pypdf==5.1.0
PyPika==0.48.9
pypng==0.20220715.0
pyproject_hooks==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.12
python-oxmsg==0.0.1
python-socketio==5.11.1
python-telegram-bot==13.15
pytz==2023.4
PyYAML==6.0.1
qrcode==7.4.2
RapidFuzz==3.10.1
referencing==0.37.0
regex==2024.7.24
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.7.0
rpds-py==0.29.0
rsa==4.9
ruff==0.7.3
s3transfer==0.8.2
safehttpx==0.1.1
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
send-mail==0.0.2
setuptools==69.0.2
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speaklater3==1.4
SQLAlchemy==2.0.27
sqlparse==0.4.4
srsly==2.4.8
sshtunnel==0.4.0
starlette==0.41.2
stevedore==5.1.0
sympy==1.13.1
telegram-django==1.1.2
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.3.1
torchvision==0.18.1
tornado==6.1
tqdm==4.66.4
translate==3.6.1
treenode==0.1.5
typed==0.6
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
ua-parser==0.18.0
unstructured==0.16.5
unstructured-client==0.27.0
uritemplate==4.1.1
urllib3==1.26.18
user-agents==2.2.0
uvicorn==0.30.3
uvloop==0.19.0
virtualenv==20.25.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.0.0
wasabi==1.1.3
watchfiles==0.22.0
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==2.3.8
wheel==0.41.3
whitenoise==6.7.0
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.19.2
